Player, RacingPlayers 리네이밍
