+++ src/main/kotlin/racingCarMVC/controller/RacingController.kt
@@ -2,7 +2,7 @@ package racingCarMVC.controller
 
 import racingCarMVC.domain.RacingService
 import racingCarMVC.domain.dto.InputData
-import racingCarMVC.domain.RacingPlayers
+import racingCarMVC.domain.RacingGameResult
 import racingCarMVC.view.InputIO
 import racingCarMVC.view.InputMessage
 import racingCarMVC.view.InputView
@@ -26,7 +26,7 @@ class RacingController {
 
     fun racingStart(inputData: InputData) {
         val racingResult = racingService.racingStart(inputData)
-        val winners = RacingPlayers(racingResult.last().players).getWinners()
+        val winners = RacingGameResult(racingResult.last().gameRecord).getWinners()
 
         resultView.showRacingResult(racingResult, winners)
     }
