+++ src/main/kotlin/racingCarMVC/domain/dto/InputData.kt
